Refactored the entire usage floats to doubles for easier and better math calculations

+++ src/core/Main.java
@@ -5,8 +5,8 @@ import core.math.Color;
 import core.math.Vector3;
 import core.shapes.Shape;
 import core.shapes.Sphere;
-import core.world.Ray;
 import core.world.Light;
+import core.world.Ray;
 
 import javax.imageio.ImageIO;
 import javax.swing.*;
@@ -21,7 +21,7 @@ public class Main
     public static void main(String[] args)
     {
         outputRenderedImage = new BufferedImage(WIDTH, HEIGHT, BufferedImage.TYPE_INT_RGB);
-
+        
         Sphere mySphere2 = new Sphere(200.0f, 200.0f, 100.0f, 100.0f, new Color(0, 255, 0));
         //core.shapes.Sphere mySphere3 = new core.shapes.Sphere(150.0f, 150.0f, 50.0f, 50.0f, new core.math.CColor(0, 255, 0, 255));
         shapeList.add(mySphere2);
@@ -34,7 +34,7 @@ public class Main
         {
             for (int x = 0; x < WIDTH; x++)
             {
-                Ray myRay = new Ray(new Vector3((float) x, (float) y, -30.0f), new Vector3(0, 0, 1));
+                Ray myRay = new Ray(new Vector3((double) x, (double) y, -30.0f), new Vector3(0, 0, 1));
                 for(Shape shape : shapeList)
                 {
                     if(shape.intersects(myRay, 1))
